+++ lib/redis/index.ts
@@ -1,2 +1,0 @@
-export { RedisCache } from './RedisCache.js'
-export type { RedisCacheConfiguration } from './AbstractRedisCache.js'